fix: address review comments for Forge-6321
